merged: changes from trunk. revs: r29311:29465
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing2@29466 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libshout/list.h
@@ -192,7 +192,7 @@ static inline void list_splice_init(struct list_head *list,
  * @member:	the name of the list_struct within the struct.
  */
 #define list_entry(ptr, type, member) \
-	((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+	((type *)((char *)(ptr)-(uintptr_t)(&((type *)0)->member)))
 
 /**
  * list_for_each	-	iterate over a list
